(#13) Hierarchy includes sub-hierarchy
Any time a hierarchy has multiple levels as directories, ie
 #{country}/#{dc}/#{fqdn} and one of the levels is not defined, we do not
want to process this in hiera. If any level of one line is undefined,
that line will either end with /, begin with /, or have // in the middle
somewhere, and we don't want to look up data files for that.

This regex just checks for that before yielding to the backend.