Fixes: Hiera errors on empty yaml files
If a completely empty yaml file is loaded then it will return `false`,
but if a yaml file is loaded with only '---' in it, it returns `nil`.
In either case hiera should handle the cases as expected instead of
erroring. This commit causes these answers to be considered the same as
if the file didn't exist at all.