Test cases for #21
Previously we weren't actually testing *anything* having to do with
*actual* yaml loading. So the previous test cases were writing what we
expected it to be, then checking that it was what we expected.

Using `YAML.load` I am simulating loading actual yaml files for the
tests, which should match the expected return values, not just expected
matching expected.

For the `hiera.yaml` config, we didn't test behaviour if it was missing
or empty.