Change Rakefile to autoversion
This change modifies the Rakefile to derrive the version of the package
based on a call to "git describe." Also adds a tasks directory for
additions to the Rake tasks.

git describe will show the current annotated tag or the
tag-commits-gSHA1 as the version string. Due to the way gem processes
version strings, the Rakefile replaces '-' with '.' We are also
stripping the gSHA1 portion, as it causes tests failures with alphabetic
characters in the version string. The version number is still uniq based
on the git-describe call.

So, the new release process would be:
  tag (signed), rake default

Future enhancements may add a task for uploads to the puppetlabs
downloads area.

Note: to test I tagged the first commit as 0.0.1.
Other options would be to retag the released versions with signed
tags.

Signed-off-by: Michael Stahnke <stahnma@puppetlabs.com>