Parse hiera.yaml at load-time
Before this commit `hiera.yaml` is loaded as yaml at runtime in an
atomic operation, but I desired to be able to take scope into account
for the purpose of adding yaml refereces and anchors.

With this commit, Hiera will read the contents of the `hiera.yaml`
config file, parse the contents for `%{strings}`, and then parse the
resulting file from yaml into ruby datastructures.

This commit is a bit of a refactor. Here is what I did to accomplish
this:
* Move `parse_string` from `Hiera::Backend` to `Hiera`
  * This method is used by both `Hiera::Config` and `Hiera::Backend`
* Allow `scope` to be passed into the `Hiera` constructor
* Use the `scope` when loading `hiera.yaml`

All tests pass.