(#13641) Fix Hiera::Backend#parse_string to support :undefined in extra_data
The method #parse_string supports an extra field 'extra_data' which is used by
hiera-mysql. When combined with hiera-puppet, sometimes the symbol :undefined
is returned, to signify that the data is undefined in Puppet.

While we do have tests and code to capture this state, the method does not
support falling back to 'extra_data' in this situation.

This patch supports the proper fallback to extra_data if it sees the special
symbol :undefined, so that these backends can be properly combined.