(#14147) Handle lack of rubygems gracefully
Hiera previously expected rubygems to be available. When running from source or
a package, rubygems may not be available. This fix rescues the load error on
rubygem and silently ignores it so that if rubygems is present, it can be used,
but if it isn't present, hiera can still function.