(maint) Fix failing spec test on Ruby 1.9.3
Without this patch the specs tests do no pass on Ruby 1.9. The output
for `Load Errors` has changed between Ruby 1.8.7 and 1.9.3. This causes
the `Hiera#logger` spec to fail on Ruby 1.9.3 as we are relying on the
output from 1.8.7.

Instead of relying on a string which is subject to change in the future,
we are now checking that the correct logger is being used.