(#14256) Adding save and delete APIs
This patch adds new APIs for saving and deleting data by delegating to a
specific backend. The backend is responsible for saving or deleting data
based on a key, value, and source. The backend must return a status response
of true or false depending on the success or failure of the save/delete
operation.

Example:

If we have the following Hiera configuration:

   :hierarchy:
     - %{environment}
     - common

   :backends:
     - backend_with_save_support

We can save data to the `backend_with_save_support` like this:

    require 'hiera'

    hiera = Hiera.new
    status = hiera.save('key, 'value', 'backend_with_save_support', 'common')

We can delete data from the `backend_with_save_support` like this:

    require 'hiera'

    hiera = Hiera.new
    status = hiera.delete('key, 'value', 'backend_with_save_support', 'common')

This patch includes tests.