(#13600) Monkey patch mktmpdir on ruby 1.8.5
The yaml_backend spec test uses mktmpdir in its tests, which is not available
on ruby 1.8.5. This fix defines mktmpdir using the 1.8.7 method if the method
is not available, as is the case in ruby 1.8.5. It defines the method if needed
in the spec_helper file, and the requires in the yaml_backend_spec are
reordered to ensure that spec_helper is required last, so that it will know if
mktmpdir needs to be defined.