Update mac packaging to dynamically generate preflight
To avoid collisions between preflight files and to avoid having to maintain
discrete lists of files, this commit moves the preflight file to an erb
template that is generated on the fly during package creation. It only adds
files to the removal list, as we have no guarantees on who owns what
directories.