(#12122) Correctly fall through backends during loop
When calling lookup for hash or array types Hiera would incorrectly return an
empty result from a backend that did not find a value instead of checking the
next backend. This happened because the empty hash and empty array were used
internally to represent a failed lookup. This changes it to use nil
everywhere. A distinct sentinel value would be ideal, but since Hiera
currently has no way to communicate to the outside world the difference
between lookup up nil and a failed lookup, it is sufficient to use nil internally.