Rework packaging to use packaging repo
This commit takes the hiera-specific data for
packaging files and places it in ext, and removes
much of the built-in packaging from hiera. A
package:bootstrap task is added which pulls in
the packaging repo, which contains packaging
tasks that are in the load path of the
top-level Rakefile. A package:teardown task
removes the packaging repo. yaml files are
added specifying packaging details for use
with the packaging repo.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>