Add json backend to hiera.
This commit brings the json backend and corresponding spec test from hiera-json
into hiera. The spec test has been updated to move some of its requires into the
spec_helper in a previous commit. This commit also updates the project_data.yaml
to add a json dependency to hiera to the gem, deb and rpm packages.