(maint) Fixup hiera.spec for el5 oddities, rpmlint issues
This commit updates hiera.spec to remove all but one of the rpmlint
warnings/errors that were being thrown. It marks /etc/hiera.yaml as a
configfile, sets _sharedstatedir to something sane on el5, updates the license
to the correct shorthand, and tweaks the summary. One lint error remains, which
is a lack of manpage for the hiera binary.