(maint) Remove CHANGELOG from Hiera
This commit removes the CHANGELOG file in the top level
of Hiera, which only contains commit hashes/ messages
from the previous releases, and is of debatable utility.
This information can be gleaned easily from a git log,
and is not of much use without git.

It also updates the packaging to remove references to the CHANGELOG.