Only load rubygems in bin/hiera
We should not require rubygems when used as a library. When invoked from the
command line we require rubygems if bundler has not been loaded. If bundler
has been loaded we assume rubygems is already loaded if it needs to be, as
requiring rubygems after loading bundler can break the bundler setup.