(Maint) Describe desired behavior in backend specs
The names used for the previous tests of the Backend code were very
tightly tied to the implementation and did very little to describe the
desired behavior. This tries to reword them in a manner that makes the
behavior that is expected more clear. It also removes the constant use
of the word "should" in favor of more specific and active verbs.