(#17434) Simplify logic around looking up values
This inverts the if statement used for looking in the scope or the
extra_data. At the same time it uncovered a bug, and fixes it, whereby
if the extra_data contained boolean false, it would be ignored. The new
behavior is that a false value in the extra_data is used rather than
changed to nil.