Add install.rb to hiera
Previously installing hiera from source was different from Facter and Puppet.
This commit aims to change that by adding an install.rb to hiera. It has many
of the same flags as the puppet and facter install.rb, notably lacking
deprecated and unused flags. This allows a user hiera to run `ruby install.rb`
inside of hiera and have the libs added to their sitelibdir, a default conf
file laid down, and the binary installed.