(maint) Fix gemspec version
Without this patch Hiera cannot easily be used from Git source to
satisfy the dependency other projects like Puppet have on Hiera.  This
is a problem because we need the means to integrate the Hiera master
branch with the Puppet master branch in our CI system.

This patch addresses the problem by adding the same public version API
that Puppet and Facter have.  In addition, a gemspec file is added to
the root of the repository in order to make a plain git checkout usable
with Bundler.