Update example hiera.yaml for Psych parser
Bare %{clientcert} in yaml is invalid, and the Psych parser will not load such
a file. The Syck parser happily accepted such files. The solution is to quote
the %{clientcert} so it becomes "%{clientcert}" and Psych recognized it as a
string. Relevant part of the YAML spec: http://yaml.org/spec/current.html#%
directive/