Update for new osx packaging target
This commit ports work from @ccaviness in
https://github.com/puppetlabs/facter/pull/512 and @djmitche in
https://github.com/puppetlabs/facter/pull/517/files to hiera, to update the osx
packaging for the soon-to-be new osx target install location of
/usr/lib/ruby/site_ruby (/Library/Ruby/Site). In order to ensure Puppet Labs
projects are ready for OSX 10.9, we have to get out of the hardcoded 1.8
directory. In order to transition smoothly, we have to blow away the old
installs and move them to the new one.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>