(22329) Only delete hiera files in the preflight
Previously, the OSX preflight removed the contents of that ‘hiera’ directory,
which also happens to be where the puppet.pkg places the contents of
hiera_puppet. The fix is simple – from the hiera dir, only remove the files
that exist in the ‘hiera’ project, not those in puppet. This commit
accomplishes this by only deleting files, not the directories. It's a bit
crufty, but its better than clobbering the wrong files.

Signed-off-by: Moses Mendoza <moses@puppetlabs.com>