Revert "Merge pull request #137 from lollipopman/add_hiera_lookup"
Errors raised in https://jenkins.puppetlabs.com/job/Puppet%20Specs%20(stable)/258/
are due to the fact that this doesn't handle the old style interpolation
correctly. This can't be fixed before the ehend of the day so this has
to be rolled back.

This reverts commit a68a9c77e988c7a50eea17dabc2b1a738bf5a220, reversing
changes made to 256970c11db5c3af0c362378973c710495b128c7.