(packaging) (#22905) rm quilt from deb packaging
We've added Saucy to the list of distrobutions we support, and it seems
that deb helper has changed slightly with this release. As a result,
building hiera packages for saucy doesn't work. This is because we
require quilt even though we have no patches that we're including. In
order to make sure that we can build hiera on saucy, we need to git rid
of that call to quilt.