(HI-186) Express Windows specific gem dependencies
Previously, if a ruby module on windows tried to use hiera as a library,
the calling module had to know about hiera's platform specific
dependencies. This comes up when trying to develop and test puppet
modules on windows.

This commit adds the windows specific gems to project_data.yaml, and
updates the Gemfile to add these dependencies when during
`bundle install`

The build pipeline will automatically create an x86-mingw32 gem in
addition to the generic one.