(HI-179) Simplify and Rakeify acceptance tests
The hiera acceptance tests didn't use the puppet master nor server, they
only used puppet apply to setup the system. This means that we can use
half the number of VMs when running tests. It also means that there
isn't any need for much of the setup procedure that puppet uses.

In addition there is now a Rakefile and task for running the beaker
tests and a Gemfile to pull beaker in. It is modelled after what is done
in Puppet but doesn't copy everything there. We may bring more over in
time.