(HI-238) Remove current directory from Ruby load path.
The current directory ('.') is on the load path for Ruby 1.8.7.
This is a security vulnerability as it allows arbitrary code loading if
users create ruby source files with names that correspond to those that
hiera is trying to load.

The fix is to explicitly remove '.' from the load path before any code
is loaded by hiera.