(maint) Fix rake spec tests
Currently the Rakefile is setup to load files for building the release
packages but is erroring out before `rubygems` and `rspec/core/raketask` are
being loaded causing the rake spec tasks to not be created only if the
packaging repo has not been bootstrapped. This prevents you from running
`rake spec` in a bare repository.

This commit changes the loading order so that rake will gracefully load
the file and create the rake spec tasks regardless of if the packaging
repo has been bootstrapped or not.