(HI-304) Make order override work with interpolation
The existing implementation of hiera interpolation did not utilize the
order override parameter of lookup. If an interpolated variable was
expected to be looked up in the overridden datasource it would instead
only be looked up in the configured datasources.

This patch passes the order_override parameter down the call stack to
the hiera_interpolate method of the Hiera::Interpolate class. This
method then in turn passes the order_override parameter back into the
Hiera::Backend lookup method. Thereby allowing the overridden datasource to
be searched for lookup values.