(HI-306) Break options parsing away from Mcollective scope load
The Hiera CLI uses ruby's optparse module to pull options out of ARGV. So
does Mcollective. The problem is that the CLI parses some options, comes
across a need for mcollective, then calls it immediately before the rest of
ARGV is parsed. That creates some weird results.  Specifically, it adds an
ordering requirement between -c and -m on the command line that is non-obvious
and undocumented.  This patch moves scope loading into a later code chunk,
after optparse is done with what it's doing.