(maint) Normalize Hiera Testing with AIO Additions
Previously Hiera testing was a bit unloved, only testing out of git and
using none of the conventions utilized by Facter or Puppet. This brings
over the Rake tasks and configs of Puppet and Facter and moves
acceptance_tests -> acceptance for consistancy (there's a symlink for
backwards compatability). There is now a git task that runs in the same
manner as Puppet/Facter as well as a task that should work for running
package tests (if there are Hiera packages). There are also the
requisite changes needed for AIO testing.