(HI-327) Set acceptance tests to run on agent role
This commit changes the node role that the acceptance tests run on from
`master` to `agent`.

This change is required in order for the tests to run test configurations
using the All In One agent packaging. The AIO testing applies directly to
agent nodes, and since Hiera functionality does not require a master, pinning
the test execution to the master role should be generalized to the agent.