(HI-14) Allow numeric key segment to identify array element
Enhances that lookup algorithm so that a key segment in the form
of an integer can be used as an index in an array. It is required
that the instance to be indexed by such a segment is an array (or nil).

This commit also includes some minor fixes in response to PR review
comments.