(HI-337) Distinguish between nil values and missing keys
This commit changes the JSON and YAML backends so that they throws
a :no_such_key exception when the key is missing and returns nil for
a key that is found but has a nil value. The Backend that calls those
backends will recognize this distinction and only use the 'extras'
hash and 'default' value when the key is missing. In contrast, a nil
return will be treated as a valid value and returned to the caller.