Don't merge: This adds explain functionality
This adds the explain option to describe how hiera would resolve
datasources. It re-orders argument parsing slightly so this can be run
on its own, or can be output prior to the data resolution being
displayed. Context is that people find hiera lookups fiendishly
non-intuitive, so having it rendered in plain text would hopefully make
the process more consumable.

Please don't merge as-is. I'd like feedback as to whether this line of
thought would be accepted into core rather than as an external script
like we use in the classroom. If so, I'll finish it and with tests and
better error handling and a puppet DSL function and the like.