(maint) Add deep_merge gem install to git and packages pre-suite
Puppet 4.0 requires the deep_merge gem. It needs to be available when
running the acceptance suite on git or packages against Puppet (it is
included in aio).

This adds a 'gem install deep_merge' on all hosts (for the sake of
simplicity) except windows hosts (where it is not needed).