(HI-348) Enable deep-merge options to be passed in lookup
Prior to this commit it was not possible to provide deep merge options
in each call. The only way to use such options was to configure them
globally in the Hiera config. This commit changes this by allowing
the 'resolution_type' parameter to be a Hash and also ensure that is
propagated to the Backend::merge_answer() method.

When the 'resolution_type' is a hash, the merge_answer method will
completely disregard the Hiera deep merge config and instead pick the
merge behavior (keyed by :behavior) from the hash and consider the
deep merge options to the remaining top level entries of the hash.