(maint) Load packaging and spec tasks separately
When building packaging on jenkins, we don't have the rake library
available. Because of this, we try to load it in our Rakefile, it fails,
and then it bombs out and doesn't load anything else (aka the rake tasks
we use for packaging).

Since we also need to load rubygems and rspec/core/rake_task when
running `rake spec`. We need to be able to run spec tests in a repo
where the packaging repo hasn't been cloned and packaging.rake isn't
available. So we need to load these two different sets of libraries
separately.