(HI-374) configure hiera to work out of the box
Prior to this commit, users had to configure hiera.yaml
before being able to use Hiera with Puppet. This commit
defines a datadir for *nix and windows systems by default
that will work with Puppet while allowing the user
to override the setting in hiera.yaml. This commit also
changes the included hiera.yaml to specify a useful
hierarchy.

please see PE-894