(maint) Improve error message in read_file for filecache
Previously if a type error occurred during the file_read, the error
message would always say that the data retrieved is a string, and not
the expected type. This is because data comes from File.read, which
returns a string. Having the class of the data from the @cache object
allows the error message to be more helpful and return the class of the
parsed data instead of just 'String'.