(HI-471) Allow backends to perform subkey lookup
Previously, Hiera performed subkey lookup by asking the backend for the
top-level key and descending into the data.  As an optimization, allow
the backends to handle the subkeys so that it can minimize processing
required to return a large data structure when only a small portion of
it is used.  This can speed up lookups for backends where processing
data is expensive (like eyaml or custom database backends).

Backends can provide a new method called lookup_with_segments.  If the
backend responds to this method, it will be called instead of lookup.