(HI-220) Deprecate use of interpolation methods in Hiera config
Using interpolation methods in the hiera configuration configuration
file should be avoided. The 'hiera' method in particular will lead
to problems since it's very likely that it will cause an endless
recursion when a new attempt is made to read the configuration file.

This commit ensures that a deprecation warning is issued when the
interpolation method syntax is used in a hiera configuration file. An
error will be raised in case the 'hiera' interpolation method is used.