(HI-486) Allow host configs to be generated at runtime
This commit adds a dependency on beaker-hostgenerator in the
acceptance Gemfile, which may be overridden using the
`BEAKER_HOSTGENERATOR_VERSION` environment variable.

The acceptance Rakefile will continue to use the `CONFIG` environment
variable, which points to an existing host config file. If `CONFIG` is
not specified, but `PLATFORMS` is, then the Rakefile will use
beaker-hostgenerator to dynamically generate a host config. The host
config is generated in the `hiera/acceptance/tmp` directory, as
opposed to `/tmp` so that it doesn't conflict with concurrent
acceptance tests running on the same CI coordinator. It also appends
SecureRandom to the hostfile, because it seems like a good idea. The
resulting host config is copied to the `log/latest/config.yml`.

For example, to run against a tagged build:

    $ bundle exec rake ci:test:aio PLATFORMS=redhat7-64a SHA=1.3.2

To run against a non-tagged commit:

    $ bundle exec rake ci:test:aio PLATFORMS=redhat7-64a \
      SHA=<full sha> SUITE_VERSION=<git describe>

Conflicts:
	acceptance/Gemfile