(HI-943) Update hiera.bat to current standard
With the switch over to building puppet-agent for windows with Vanagon,
we need to effectively install hiera with install.rb. This script
installs hiera.bat, which is a wrapper that lets windows handle ruby
scripts. This commit updates hiera.bat to use our environment.bat file if
we install it, and to be more in line with the hiera.bat file installed
with the puppet-agent package as built by puppet_for_the_win