(HI-494) Bring back support for empty interpolation
Before this, we started doing something wrong in the hiera 3.x series as
interpolation of empty strings - i.e:

* %{}
* %{ }
* %%{}

No longer worked. As a side effect of the changd behavior it was intead
possible to do this:

* %{::}

With this commit the earlier behavior is supported, and these cases now
all produce an empty string:

* %{}
* %{ }
* %{::}
* %{ :: }

The case of including an interpolation in the result is also supported
* %%{}{environment}

This produces a result with the value "%{environment}"

This also adds unit testing of interpolation. What is sad is that the 6
added examples constitutes an 300% increase in the number of
interpolation tests.