(HI-499) Fail if backend cannot be loaded
Before this, if a backend could not be loaded, this was only logged as a
warning and hiera merrily continued by serving data from remaining (if
any) backends.

This can lead to catastrophic results as lookups may be resolved with
large portions of the data missing. This in turn can cause catalogs to
contain a radically different configuration causing a wide range of
unpleasant surprises.