(HI-504) Add ability to use quoted keys
This commit adds the ability to put quotes around keys. A key can be
quoted with either single or double quotes. A quoted key will be used
verbatim instead of splitting it on dots. In other words, quoting a key
will not use dots to navigate into hashes and arrays. Instead it will
match a dotted key at top level.

Interpolation expressions that use interpolation methods must have the
key quoted twice to get the effect. The reason for this is that the
argument to an interpolation method already had to be quoted. In order
to not introduce incompatibilities, a quoted key in this context must
therefore be quoted twice. Example:

```
---
a.quoted.key: 'Value for a quoted key'
ref.to.key: 'Put the value here: %{hiera("''a.quoted.key''")}'
```