(HI-508) Prevent that subkeys can do lookups in string values
Before this commit, a subkey of a dotted key could be matched by a
string that contained a matching text. Hiera would require that the
value
responded to :[] and that it was not an Array (arrays should use numeric
index). A string fulfills that requirement.

This commit adds the constraint that the value cannot be a String when
subjected to a lookup.