(HI-510) Look for hiera.yaml in codedir first, then confdir
This commit ensures that when Hiera is initialized, it will first check
if a config file is specified using a CLI option. If it is, then that
config will be used (and an error is raised if it doesn't exist). When
no CLI option is used, Hiera will first look `hiera.yaml` in the
codedir and if no such file exists, it will default to `hiera.yaml`
in confdir.