(HI-544) Fail with meaningful message on hiera configs with version > 3
This commit adds a check for a `version` field in the hiera.yaml config
file. If the version is present and indicates a number greater than 3, then
the load fails with an error.

This is a cherry-pick of fd10d3f16b0f319af817ea494ae768ea7e350a32 to
reapply the patch following a mistaken revert.