(HI-596) Move hiera to running test from beaker-puppet
This includes removing all the set up logic from hiera. This is now
stored in beaker-puppet. There is one test that was previously being
executed during the test phase, but is more appropriate to run as a
pre-suite. This ensures it is run before any of the other tests, so that
order isn't as important.